+++ src/views/DestinationDetailView.vue
@@ -17,6 +17,7 @@ import UserTravelStories from '../components/UserTravelStories.vue';
 import AudioGuidePlayer from '../components/AudioGuidePlayer.vue';
 import TravelExpensesExporter from '../components/TravelExpensesExporter.vue';
 import DestinationComparator from '../components/DestinationComparator.vue';
+import CurrencyCalculator from '../components/CurrencyCalculator.vue';
 
 const route = useRoute();
 const travelStore = useTravelStore();
@@ -64,6 +65,7 @@ const relatedPosts = computed(() => {
         <UserTravelStories :destination-id="destination.id" class="margin-top-calc" />
         <TravelMapVisualizer :highlight-dest-id="destination.id" class="margin-top-calc" />
         <DestinationComparator class="margin-top-calc" />
+        <CurrencyCalculator class="margin-top-calc" />
         <TravelExpensesExporter :destination-id="destination.id" class="margin-top-calc" />
         <PackingChecklist class="margin-top-calc" />
       </div>